requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
entrypoints==0.3
ffmpeg-python==0.2.0
flake8==3.7.9
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
moviepy==1.0.1
numpy==1.17.4
pathspec==0.6.0
Pillow==6.2.1
proglog==0.1.9
pycodestyle==2.5.0
pyflakes==2.1.1
pygifsicle==1.0.1
pylint==2.4.4
regex==2019.11.1
requests==2.22.0
six==1.13.0
sty==1.0.0b12
toml==0.10.0
tqdm==4.40.0
typed-ast==1.4.0
urllib3==1.25.7
Wand==0.5.7
wrapt==1.11.2
